+++ src/konnektor/network_planners/concatenators/__init__.py
