+++ src/konnektor/network_planners/concatenators/_abstract_network_concatenator.py
