+++ src/konnektor/network_planners/concatenators/cyclic_concatenator.py
@@ -15,7 +15,7 @@ class MstConcatenate(NetworkConcatenator):
     def __init__(self, mapper: AtomMapper, scorer: AtomMappingScorer, n_connecting_edges: int = 3,
                  node_present_in_cycles: int = 2, cycle_sizes: Union[int, List[int]] = 3, nprocesses: int = 1):
         """
-        This concatenator is connnecting two Networks with a kruskal like approach up to the number of connecting edges.
+        This concatenators is connnecting two Networks with a kruskal like approach up to the number of connecting edges.
 
         Parameters
         ----------
